Fixed compilation warning.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@59188 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/corelib/expr.cpp
@@ -655,7 +655,7 @@ CExprParser::Scan(bool operand)
             m_VStack[m_v_sp].SetType(CExprValue::eSTRING);
 
             m_VStack[m_v_sp].m_sval = StringConstant;     //assign string
-            m_Pos += pos+1;                               // move m_Pos
+            m_Pos += (int)pos+1;                          // move m_Pos
 
             m_VStack[m_v_sp].m_Pos = m_Pos;
             m_VStack[m_v_sp].m_Var = NULL;
